refactor: modularize and remove question name from rancher compose
