+++ processRancherCompose.go
@@ -56,24 +56,15 @@ func processRancherCompose(rancherCompose map[interface{}]interface{}) {
 	}
 
 	firstQuestion := make(map[string]interface{})
-	firstQuestion["variable"] = "NAME"
-	firstQuestion["default"] = catalog["name"].(string)
-	firstQuestion["label"] = "Kubernetes Name (Max 24 characters)"
-	firstQuestion["description"] = "at most 24 characters] = matching regex [a-z]([-a-z0-9]*[a-z0-9])?)"
+	firstQuestion["variable"] = "NAMESPACE"
+	firstQuestion["default"] = "default"
+	firstQuestion["label"] = "Kubernetes Namespace"
+	firstQuestion["description"] = "Make sure the Namespace exists or you will not be able to see the service"
 	firstQuestion["required"] = true
 	firstQuestion["type"] = "string"
 
-	secondQuestion := make(map[string]interface{})
-	secondQuestion["variable"] = "NAMESPACE"
-	secondQuestion["default"] = "default"
-	secondQuestion["label"] = "Kubernetes Namespace"
-	secondQuestion["description"] = "Make sure the Namespace exists or you will not be able to see the service"
-	secondQuestion["required"] = true
-	secondQuestion["type"] = "string"
-
-	newQuestionsArray := make([]interface{}, 2)
+	newQuestionsArray := make([]interface{}, 1)
 	newQuestionsArray[0] = firstQuestion
-	newQuestionsArray[1] = secondQuestion
 
 	newQuestions := Append(newQuestionsArray, questions...) // The '...' is essential!
 	catalog["questions"] = newQuestions
